+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add generic FileUpload mutation - #6470 by @IKarbowiak
 - Multichannel MVP: Multicurrency - #6242 by @fowczarek @d-wysocki
 - Fix password reset request - #6351 by @Manfred-Madelaine-pro, Ambroise and Pierre
+- Add possibility to exclude products from shipping method - #6506 by @korycins
 
 # 2.11.1